requirements.txt
selenium
chromedriver-py
python-dotenv
pandas